Fixed a bug when unicode content is downloaded, was not misinterpreted as such
When a Unicode web page is downloaded, it goes into a file, and this
was loaded with no explicit encoding set.  Result: characters with
surrogate code points, like the pound sign, get turned into several
garbage characters.  By detecting the encoding and loading the file
accordingly, this means the resulting HTML has the correct entities.